refactor: reorganize alert constants and simplify imports in DialAlert component
